Fixes pathing for Windows bundled EXE.
When the app is combined with nw.exe into a single file, it has to
be extracted to a temp dir when it's later executed. The temp dir
will be the CWD for the process and will contain the contents of
the zipped source tree. Unfortunately, this makes it more difficult
to find files packaged alongside the executable (as with the games/
directory).

This change fixes all broken paths when running from a temp dir
created by a single-file executable. The deployed directory tree
should look like:

moonshot.exe
games/
games/fearless_night/
games/akbb/
...